tests(socialaccount): Refactor tests.py -- split into multiple files
